list_next bug
Fixed a bug in release_lock's iteration of the lock_list. When a lock
is removed, the current element should walk backwards by one, not forwards,
since the for loop will walk it forwards anyway.

The bug does not happen as often anymore (alarm-negative, e.g., now passes).

Also, cleaned up coding style in lock_release.

TODO: The priorty-change still causes list_next to trigger an assert. I was
not able to trace the problem to its root, but, after playing around with GDB,
it looks like the lock_list is eventually corrupted. A list element that is
NOT the tail ends up with a next pointer holding NULL, which causes the assert
to fail (the list element is not the head, nor is it an interior element). It's
possible that some of the code that I added to thread.c caused this problem, but
I can't see how it would.